+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Added Unit Tests for RouteSearchParameters.class() (while fixing Issue #291)
+- Added ability to turn warning messages in the route response which can be used for showing info to a user when warning criteria have been met based on extended storages.
 ### Fixed
 - If residential penalty reduces speed to <5, set it to 5
 - Added a new ParameterValueException in RouteSearchParameters if the profile is driving-car and profile_params are set in the options (Issue #291)